[flex-box] renamed folder
