+++ semcore/flex-box/src/invalid-state-box/InvalidStateBox.tsx
