+++ semcore/flex-box/src/invalid-state-box/invalidStateBox.shadow.css
